requirements.txt
Django==1.8
PyYAML==3.11
Whoosh==2.6.0
## FIXME: could not find svn URL in dependency_links for this package:
bulbs==0.3.29-20150324
distribute==0.7.3
django-bootstrap3==5.2.0
django-haystack==2.3.1
django-taggit==0.13.0
httplib2==0.9
ipython==3.0.0
omnijson==0.1.2
pep8==1.6.2
python-dateutil==1.5
requests==2.6.0
six==1.9.0
wsgiref==0.1.2
